chore: split decimal cast
